#130 - add packages and restructure
