+++ src/main/java/com/artipie/maven/asto/RepositoryChecksums.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artipie.maven.repository;
+package com.artipie.maven.asto;
 
 import com.artipie.asto.Key;
 import com.artipie.asto.Storage;
