+++ src/main/java/com/artipie/maven/http/ArtifactGetResponse.java
@@ -30,7 +30,7 @@ import com.artipie.http.async.AsyncResponse;
 import com.artipie.http.rs.RsWithBody;
 import com.artipie.http.rs.RsWithHeaders;
 import com.artipie.http.rs.StandardRs;
-import com.artipie.maven.repository.RepositoryChecksums;
+import com.artipie.maven.asto.RepositoryChecksums;
 
 /**
  * Artifact {@code GET} response.
