+++ src/main/java/com/artipie/maven/http/RemoteDownloadSlice.java
@@ -34,8 +34,8 @@ import com.artipie.http.rs.RsWithBody;
 import com.artipie.http.rs.RsWithHeaders;
 import com.artipie.http.rs.RsWithStatus;
 import com.artipie.http.rs.StandardRs;
-import com.artipie.maven.repository.ArtifactNotFoundException;
-import com.artipie.maven.repository.Repository;
+import com.artipie.maven.ArtifactNotFoundException;
+import com.artipie.maven.Repository;
 import com.jcabi.log.Logger;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
