+++ src/main/java/com/artipie/maven/http/UpdateMavenSlice.java
@@ -33,9 +33,9 @@ import com.artipie.http.rq.RequestLineFrom;
 import com.artipie.http.rs.RsStatus;
 import com.artipie.http.rs.RsWithStatus;
 import com.artipie.http.slice.KeyFromPath;
-import com.artipie.maven.AstoMaven;
 import com.artipie.maven.Maven;
-import com.artipie.maven.repository.ValidUpload;
+import com.artipie.maven.ValidUpload;
+import com.artipie.maven.asto.AstoMaven;
 import java.nio.ByteBuffer;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
